feat:  refactor routes for barber and user authentication
- Refactored barber authentication routes into separate files
- Added new routes for barber business and profile management
- Removed legacy user and barber authentication routes
- Created user authentication routes for checking OTP and completing profiles
- Added user profile management routes
- Implemented a test connection script for MongoDB
