+++ models/User.model.js
@@ -1,0 +1,15 @@
+import mongoose from "mongoose";
+
+const userSchema = new mongoose.Schema({
+  firebaseUid: { type: String, required: true, unique: true },
+  phoneNumber: { type: String, required: true, unique: true },
+  firstName: { type: String, required: true },
+  lastName: { type: String, required: true },
+  gender: { type: String, enum: ["Male", "Female", "Other"], required: true },
+  dateOfBirth: { type: Date, required: true },
+  address: { type: String, required: true },
+  email: { type: String, required: true, unique: true },
+  createdAt: { type: Date, default: Date.now },
+});
+
+export default mongoose.model("User", userSchema);
