+++ models/schemas/BreakTime.schema.js
@@ -1,0 +1,8 @@
+import mongoose from "mongoose";
+
+const breakTimeSchema = new mongoose.Schema({
+  start: String,
+  end: String,
+});
+
+export default breakTimeSchema;
