+++ package.json
@@ -19,7 +19,8 @@
     "express": "^5.1.0",
     "firebase-admin": "^13.2.0",
     "jsonwebtoken": "^9.0.2",
+    "mongo": "^0.1.0",
     "mongoose": "^8.13.2",
-    "nodemon": "^3.1.9"
+    "nodemon": "^3.1.10"
   }
 }
