+++ server.js
@@ -3,11 +3,18 @@ import express from "express";
 import dotenv from "dotenv";
 import cors from "cors";
 import connectDB from "./config/dbConnect.js";
-import router from "./routes/userAuthRoutes.js";
-import router1 from "./routes/barberAuthRoutes.js";
+
+// User routes
+import userAuthRoutes from "./routes/user/auth/userAuth.routes.js";
+
+// Barber routes
+import barberAuthRoutes from "./routes/barber/auth/barberAuth.routes.js";
+import barberProfileRoutes from "./routes/barber/profile/barberProfile.routes.js";
+import barberBusinessRoutes from "./routes/barber/business/barberBusiness.routes.js";
+
+// Legacy routes (to be refactored later)
 import router2 from "./routes/serviceRoutes.js";
 import employeeRoutes from "./routes/employeeRoutes.js";
-
 
 dotenv.config();
 
@@ -20,10 +27,16 @@ connectDB();
 app.use(cors());
 app.use(express.json());
 
-app.use("/api/barberauth", router1);
-app.use("/api/userauth", router);
-app.use("/api/service", router2)
-// app.use('/api/auth', authRoutes);
+// User API routes
+app.use("/api/user/auth", userAuthRoutes);
+
+// Barber API routes  
+app.use("/api/barber/auth", barberAuthRoutes);
+app.use("/api/barber/profile", barberProfileRoutes);
+app.use("/api/barber/business", barberBusinessRoutes);
+
+// Legacy routes (to be refactored)
+app.use("/api/service", router2);
 app.use('/api/employees', employeeRoutes);
 
 app.listen(PORT, () => {
